+++ core/data/src/main/java/com/kychan/mlog/core/data/repository/SearchRepositoryImpl.kt
@@ -5,7 +5,7 @@ import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import com.kychan.mlog.core.data.mapper.toRecentSearchEntity
 import com.kychan.mlog.core.data.mediator.SearchMovieMediator
-import com.kychan.mlog.core.dataSourceLocal.room.datasource.MovieRoomDataSource
+import com.kychan.mlog.core.dataSourceLocal.room.datasource.SearchRoomDataSource
 import com.kychan.mlog.core.dataSourceLocal.room.model.RecentSearchEntity
 import com.kychan.mlog.core.dataSourceLocal.room.model.toDomain
 import com.kychan.mlog.core.dataSourceLocal.room.model.update
@@ -19,7 +19,7 @@ import javax.inject.Inject
 
 class SearchRepositoryImpl @Inject constructor(
     private val tmdbDataSource: TMDBDataSource,
-    private val movieRoomDataSource: MovieRoomDataSource
+    private val searchRoomDataSource: SearchRoomDataSource
 ): SearchRepository {
     override fun getSearch(
         language: Language,
@@ -39,20 +39,20 @@ class SearchRepositoryImpl @Inject constructor(
     }
 
     override suspend fun updateRecentSearch(text: String) {
-        movieRoomDataSource.getRecentSearch(text).firstOrNull()?.let {
-            movieRoomDataSource.updateRecentSearch(it.update())
-        }?: movieRoomDataSource.updateRecentSearch(text.toRecentSearchEntity())
+        searchRoomDataSource.getRecentSearch(text).firstOrNull()?.let {
+            searchRoomDataSource.updateRecentSearch(it.update())
+        }?: searchRoomDataSource.updateRecentSearch(text.toRecentSearchEntity())
     }
 
     override fun getRecentSearch(): Flow<List<RecentSearch>> {
-        return movieRoomDataSource.getRecentSearches().map { it.map(RecentSearchEntity::toDomain) }
+        return searchRoomDataSource.getRecentSearches().map { it.map(RecentSearchEntity::toDomain) }
     }
 
     override suspend fun deleteAllRecentSearch() {
-        movieRoomDataSource.deleteAllRecentSearch()
+        searchRoomDataSource.deleteAllRecentSearch()
     }
 
     override suspend fun deleteRecentSearch(id: Int) {
-        movieRoomDataSource.deleteRecentSearch(id)
+        searchRoomDataSource.deleteRecentSearch(id)
     }
 }
