+++ core/dataSourceLocal/room/src/main/java/com/kychan/mlog/core/dataSourceLocal/room/datasource/MovieRoomDataSource.kt
@@ -16,15 +16,4 @@ interface MovieRoomDataSource {
         syncLogType: SyncLogType,
         currentKey: Int,
     )
-    suspend fun updateRecentSearch(
-        recentSearchEntity: RecentSearchEntity,
-    )
-
-    fun getRecentSearches(): Flow<List<RecentSearchEntity>>
-
-    fun getRecentSearch(text: String): Flow<RecentSearchEntity>
-
-    suspend fun deleteAllRecentSearch()
-
-    suspend fun deleteRecentSearch(id: Int)
 }
